requirements.txt
appdirs==1.4.3
defusedxml==0.5.0
Django==1.11.1
django-allauth==0.32.0
django-extensions==1.7.9
django-filebrowser==3.7.2
django-grappelli==2.8.3
django-hitcount==1.2.2
django-map-widgets==0.1.8
django-mptt==0.8.7
django-taggit==0.22.1
django-webpack-loader==0.5.0
docutils==0.13.1
oauthlib==2.0.2
olefile==0.44
packaging==16.8
Pillow==4.1.1
psycopg2==2.7.1
pyparsing==2.2.0
python3-openid==3.1.0
pytz==2017.2
requests==2.14.2
requests-oauthlib==0.8.0
six==1.10.0
unicode-slugify==0.1.3
Unidecode==0.4.20
